tink-streaming-aead: cope with partial reads
It is valid for a Rust `std::io::Read` object to read less than the
size of the provided buffer, even when more data is available (unlike
Go's `io::ReadFull`, which only does a partial read if EOF is truly
encountered).  Cope with this, and explicitly test for it too.

Fixes #68