signature: force use of trait method
The `ed25519::Signature` type implements `signature::Signature`,
which has a `from_bytes()` method, used in ed25519_verifier.rs.

Newer version of the `ed25519` crate also add an inherent
`ed25519::Signature::from_bytes()` method (which the trait impl now
defers to), resulting in a lint warning that the import of the
`signature::Signature` trait is unnecessary.

However, removing the trait import would break builds with an older
version of the `ed25519` dep, as the inherent method was only added
in ed25519 >= 1.3.

So force use of the trait variant of `from_bytes`.