examples: show how to build for Wasm
Tink requires cryptographic randomness, and this is provided by the
`getrandom` crate. By default, that crate does not include any support
for Wasm, but the `js` feature turns on support for randomness via WASI.

Modify one of the examples to demonstrate how to build for Wasm by
ensuring that `getrandom/js` is in the Cargo build graph.

Check this with a CI step.